@@ -7,7 +7,8 @@ import { Book } from '../../shared/book';
 @Component({
   selector: 'bm-book-create',
   templateUrl: './book-create.component.html',
-  styleUrls: ['./book-create.component.css']
+  standalone: false,
+  styleUrl: './book-create.component.css'
 })
 export class BookCreateComponent {
 
