@@ -8,7 +8,8 @@ import { atLeastOneValue, isbnFormat } from '../shared/validators';
 @Component({
   selector: 'bm-book-form',
   templateUrl: './book-form.component.html',
-  styleUrls: ['./book-form.component.css']
+  standalone: false,
+  styleUrl: './book-form.component.css'
 })
 export class BookFormComponent implements OnChanges {
   @Input() book?: Book;
