@@ -5,7 +5,8 @@ import { AuthService } from './shared/auth.service';
 @Component({
   selector: 'bm-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  standalone: false,
+  styleUrl: './app.component.css'
 })
 export class AppComponent {
   constructor(public auth: AuthService) {}
