@@ -7,7 +7,8 @@ import { BookStoreService } from '../../shared/book-store.service';
 @Component({
   selector: 'bm-book-list',
   templateUrl: './book-list.component.html',
-  styleUrls: ['./book-list.component.css']
+  standalone: false,
+  styleUrl: './book-list.component.css'
 })
 export class BookListComponent {
   books$: Observable<Book[]>;
