@@ -1,7 +1,8 @@
 import { Directive, EventEmitter, HostListener, Input, Output } from '@angular/core';
 
 @Directive({
-  selector: '[bmConfirm]'
+  selector: '[bmConfirm]',
+  standalone: false
 })
 export class ConfirmDirective {
   @Input('bmConfirm') confirmText?: string;
